+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (as of version 1.4.0).
 
+## [Unreleased]
+
+### Added
+
+- Don't crash on failure to convert illustration (skip illus instead)
+
 ## [1.4.3] - 2022-06-21
 
 ### Changed